comit

+++ resources/views/admin/size_of_organization/scriptjs.blade.php
@@ -129,37 +129,59 @@
 
         $(document).on("click", "#delete", function(event) {
             var userURL = $(this).data("url");
+            var id = $(this).data("id");
 
             event.preventDefault();
-            swal({
-                title: `Are you sure you want to delete this record?`,
-                text: "If you delete this, it will be gone forever.",
-                icon: "warning",
-                buttons: true,
-                dangerMode: true,
-            }).then((willDelete) => {
-                if (willDelete) {
-                    $.ajax({
-                        headers: {
-                            "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
-                                "content"
-                            ),
-                        },
-                        method: "DELETE",
-                        url: userURL,
-                        dataType: "json",
-                        success: function(output) {
-                            if (output == true) {
 
-                                sessionStorage.setItem('showSuccessNotification', 'true');
-                                location.reload();
+            $.ajax({
+                headers: {
+                    "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
+                        "content"
+                    ),
+                },
+                method: "post",
+                url: "{{route('sizeoforganization.selectedbyuser')}}",
+                data: {
 
+                    id: id
+                },
+                dataType: "json",
+                success: function(output) {
+                    if (output == false) {
+                        errorAlert("Size Of Organization");
 
-                            } else {
-                                toastr.error("Size Of Organization don't Deleted !");
+                    } else {
+                        swal({
+                            title: `Are you sure you want to delete this record?`,
+                            text: "If you delete this, it will be gone forever.",
+                            icon: "warning",
+                            buttons: true,
+                            dangerMode: true,
+                        }).then((willDelete) => {
+                            if (willDelete) {
+
+                                $.ajax({
+                                    headers: {
+                                        "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
+                                            "content"
+                                        ),
+                                    },
+                                    method: "DELETE",
+                                    url: userURL,
+                                    dataType: "json",
+                                    success: function(output) {
+                                        if (output == true) {
+                                            sessionStorage.setItem('showSuccessNotification', 'true');
+                                            location.reload();
+
+                                        } else {
+                                            errorAlert("Size Of Organization");
+                                        }
+                                    },
+                                });
                             }
-                        },
-                    });
+                        });
+                    }
                 }
             });
         });
